Remove now unused lua.h file
