+++ code/scripting/ade.cpp
@@ -5,7 +5,13 @@
 #include "ship/ship.h"
 #include "ade_api.h"
 
-#include <memory>
+#include "scripting/api/object.h"
+#include "scripting/api/ship.h"
+#include "scripting/api/asteroid.h"
+#include "scripting/api/debris.h"
+#include "scripting/api/waypoint.h"
+#include "scripting/api/weapon.h"
+#include "scripting/api/beam.h"
 
 namespace {
 using namespace scripting;
@@ -963,6 +969,34 @@ const char *ade_get_type_string(lua_State *L, int argnum)
 		return "Unknown";
 
 	return Lua_type_names[type];
+}
+
+int ade_set_object_with_breed(lua_State *L, int obj_idx)
+{
+	using namespace scripting::api;
+
+	if(obj_idx < 0 || obj_idx >= MAX_OBJECTS)
+		return ade_set_error(L, "o", l_Object.Set(object_h()));
+
+	object *objp = &Objects[obj_idx];
+
+	switch(objp->type)
+	{
+		case OBJ_SHIP:
+			return ade_set_args(L, "o", l_Ship.Set(object_h(objp)));
+		case OBJ_ASTEROID:
+			return ade_set_args(L, "o", l_Asteroid.Set(object_h(objp)));
+		case OBJ_DEBRIS:
+			return ade_set_args(L, "o", l_Debris.Set(object_h(objp)));
+		case OBJ_WAYPOINT:
+			return ade_set_args(L, "o", l_Waypoint.Set(object_h(objp)));
+		case OBJ_WEAPON:
+			return ade_set_args(L, "o", l_Weapon.Set(object_h(objp)));
+		case OBJ_BEAM:
+			return ade_set_args(L, "o", l_Beam.Set(object_h(objp)));
+		default:
+			return ade_set_args(L, "o", l_Object.Set(object_h(objp)));
+	}
 }
 
 }
