+++ code/scripting/api/beam.cpp
@@ -9,8 +9,6 @@
 #include "subsystem.h"
 #include "mc_info.h"
 
-extern int ade_set_object_with_breed(lua_State *L, int obj_idx);
-
 namespace scripting {
 namespace api {
