+++ code/scripting/api/weapon.cpp
@@ -10,8 +10,6 @@
 #include "team.h"
 #include "mc_info.h"
 
-extern int ade_set_object_with_breed(lua_State *L, int obj_idx);
-
 namespace scripting {
 namespace api {
