+++ code/source_groups.cmake
@@ -907,7 +907,6 @@ set(file_root_scripting
 	scripting/ade_args.cpp
 	scripting/ade_args.h
 	scripting/lua.cpp
-	scripting/lua.h
 	scripting/scripting.cpp
 	scripting/scripting.h
 )
